requirements.txt
APScheduler==3.9.1
arrow==1.2.2
Django==4.0.3
djangorestframework==3.13.1
djangorestframework_simplejwt==5.1.0
firebase_admin==5.2.0
oauth2client==4.1.3
pytz==2021.3
requests==2.27.1
